+++ example/android/app/src/main/kotlin/com/example/m7_livelyness_detection_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.m7_livelyness_detection_example
+package com.m7.poc.m7_livelyness_detection_example
 
 import io.flutter.embedding.android.FlutterActivity