+++ lib/src/core/extensions/double_extension.dart
@@ -1,0 +1,9 @@
+extension M7DoubleExtension on double {
+  bool get isEven {
+    return round() % 2 == 0;
+  }
+
+  bool get isOdd {
+    return round() % 2 == 1;
+  }
+}