+++ lib/src/core/extensions/int_extension.dart
@@ -1,0 +1,9 @@
+extension M7IntExtension on int {
+  bool get isEven {
+    return this % 2 == 0;
+  }
+
+  bool get isOdd {
+    return this % 2 == 1;
+  }
+}